+++ app/src/main/java/com/nz2dev/wordtrainer/app/presentation/modules/training/overview/OverviewTrainingsView.java
@@ -13,6 +13,6 @@ public interface OverviewTrainingsView {
     void showTrainings(Collection<Training> words);
     void updateTraining(Training training);
 
-    void navigateWordTraining(int trainingId);
+    void navigateWordTraining(long trainingId);
 
 }
